third_party/WebKit/LayoutTests/fast/dom/CompositorProxy/proxy-forces-layer.html

<!DOCTYPE html>
<div id="container"></div>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
async_test(function(test) {
    if (!window.internals)
        return;
    window.addEventListener('load', function() {
        // The container should not have a layer.
        var container = document.getElementById('container');
        var layers = JSON.parse(window.internals.layerTreeAsText(document))["layers"];
        assert_equals(1, layers.length);

        // Creating a proxy should force the container to have a layer.
        var proxy = new CompositorProxy(container, ['opacity']);
        layers = JSON.parse(window.internals.layerTreeAsText(document,
            internals.LAYER_TREE_INCLUDES_COMPOSITING_REASONS))["layers"];
        assert_equals(2, layers.length);
        assert_in_array('compositorProxy', layers[1].compositingReasons);

        // Detach the element from the document. The corresponding layer should be removed.
        container.parentNode.removeChild(container);
        layers = JSON.parse(window.internals.layerTreeAsText(document))["layers"];
        assert_equals(1, layers.length);

        // Add the element back to the document. The corresponding layer should reappear.
        document.body.appendChild(container);
        layers = JSON.parse(window.internals.layerTreeAsText(document, internals.LAYER_TREE_INCLUDES_COMPOSITING_REASONS))["layers"];
        assert_equals(2, layers.length);
        assert_in_array('compositorProxy', layers[1].compositingReasons);

        // Disconnecting the proxy should also remove the layer.
        proxy.disconnect();
        layers = JSON.parse(window.internals.layerTreeAsText(document))["layers"];
        assert_equals(1, layers.length);

        test.done();
    });
}, "This test checks that creating a CompositorProxy forces the element to have a layer");
</script>
